+++ sentry/src/test/java/io/sentry/ExternalOptionsTest.kt
@@ -177,12 +177,12 @@ class ExternalOptionsTest {
 
   @Test
   fun `creates options with shutdownTimeoutMillis using external properties`() {
-    withPropertiesFile("shutdown-timeout=2000") { assertEquals(2000L, it.shutdownTimeoutMillis) }
+    withPropertiesFile("shutdown-timeout-millis=2000") { assertEquals(2000L, it.shutdownTimeoutMillis) }
   }
 
   @Test
   fun `creates options with sessionFlushTimeoutMillis using external properties`() {
-    withPropertiesFile("session-flush-timeout=2000") { assertEquals(2000L, it.sessionFlushTimeoutMillis) }
+    withPropertiesFile("session-flush-timeout-millis=2000") { assertEquals(2000L, it.sessionFlushTimeoutMillis) }
   }
 
   @Test